requirements.txt
python-dotenv
smartapi-python
pytz
pyotp
pandas
streamlit
